exclude rule_based clustering
