+++ config/trainer/callbacks/finetune.yaml
@@ -11,46 +11,29 @@
 - _target_: pytorch_lightning.callbacks.LearningRateMonitor
   logging_interval: step
 - _target_: rmind.callbacks.WandbAttentionMaskLogger
+- _target_: rmind.callbacks.WandbWaypointsLogger
+  when: on_train_batch_end
+  every_n_batch: 100
+  key: waypoints
+  crs: "EPSG:25832"
+  data:
+    image: [data, cam_front_left, 0, -1]
+    waypoints_xy_normalized: [data, waypoints/xy_normalized, 0, -1]
+    waypoints_xy: [data, waypoints/xy, 0, -1]
+    ego_xy: [data, meta/Gnss/xy, 0, -1]
+  caption:
+    input_id: [meta, input_id, 0]
+    time_stamp: [data, meta/ImageMetadata.cam_front_left/time_stamp, 0, -1]
+    frame_idx: [data, meta/ImageMetadata.cam_front_left/frame_idx, 0, -1]
 - _target_: rmind.callbacks.PredictMetricsCallback
   prediction_config:
     objectives:
       - score_l1
       - score_l1_rel
       - score_signed_error
-  cluster_fn:
-    _target_: rmind.callbacks.RuleBasedCluster
-    fields:
-      gas:
-        key: meta/VehicleMotion/gas_pedal_normalized
-        reduce: last
-      brake:
-        key: meta/VehicleMotion/brake_pedal_normalized
-        reduce: last
-      steer:
-        key: meta/VehicleMotion/steering_angle_normalized
-        reduce: last
-      dp_g:
-        key: meta/VehicleMotion/gas_pedal_normalized
-        reduce: last_diff
-      speed:
-        key: meta/VehicleMotion/speed
-        reduce: last
-    rules:
-      - name: highway
-        when: {speed: {ge: 70}}
-      - name: braking_turn
-        when: {brake: {ge: 0.02}, steer: {abs_ge: 0.05}}
-      - name: braking
-        when: {brake: {ge: 0.02}, steer: {abs_lt: 0.05}}
-      - name: cruise_turn
-        when: {gas: {ge: 0.05}, brake: {lt: 0.02}, steer: {abs_ge: 0.05}}
-      - name: acceleration
-        when: {gas: {ge: 0.05}, brake: {lt: 0.02}, dp_g: {ge: 0.01}}
-      - name: gas_release
-        when: {gas: {ge: 0.05}, brake: {lt: 0.02}, dp_g: {lt: -0.01}}
-      - name: cruise
-        when: {gas: {ge: 0.05}, brake: {lt: 0.02}, steer: {abs_lt: 0.05}, dp_g: {abs_lt: 0.01}}
-    default: idle_coast
+      - prediction_std
+      - prediction_diff_prev
+  cluster_fn: ${cluster_fn}
   # cluster_metrics:
   #   acceleration:
   #     - policy/score_l1/value/continuous/gas_pedal
@@ -70,17 +53,3 @@
   #   highway:
   #     - policy/score_l1/value/continuous/gas_pedal
   #     - policy/score_l1/value/continuous/steering_angle
-- _target_: rmind.callbacks.WandbWaypointsLogger
-  when: on_train_batch_end
-  every_n_batch: 100
-  key: waypoints
-  crs: "EPSG:25832"
-  data:
-    image: [data, cam_front_left, 0, -1]
-    waypoints_xy_normalized: [data, waypoints/xy_normalized, 0, -1]
-    waypoints_xy: [data, waypoints/xy, 0, -1]
-    ego_xy: [data, meta/Gnss/xy, 0, -1]
-  caption:
-    input_id: [meta, input_id, 0]
-    time_stamp: [data, meta/ImageMetadata.cam_front_left/time_stamp, 0, -1]
-    frame_idx: [data, meta/ImageMetadata.cam_front_left/frame_idx, 0, -1]
