+++ src/rmind/callbacks/__init__.py
@@ -7,7 +7,7 @@ from .loggers import (
     WandbWaypointsLogger,
 )
 from .logit_bias import LogitBiasSetter
-from .predict_metrics import PredictMetricsCallback, RuleBasedCluster
+from .predict_metrics import PredictMetricsCallback
 from .prediction import (
     DataFramePredictionWriter,
     RerunPredictionWriter,
@@ -23,7 +23,6 @@ __all__ = [
     "PredictMetricsCallback",
     "PredictionConfigSetter",
     "RerunPredictionWriter",
-    "RuleBasedCluster",
     "TensorDictPredictionWriter",
     "WandbAttentionMaskLogger",
     "WandbImageParamLogger",
