+++ src/rmind/utils/__init__.py
@@ -1,0 +1,3 @@
+from .cluster import RuleBasedCluster
+
+__all__ = ["RuleBasedCluster"]
